start refactoring standard library. temporarily remove from runtime to get tests to compile
