+++ runtime/convertValues_test.go
@@ -1154,7 +1154,6 @@ func TestImportRuntimeType(t *testing.T) {
 			expected: interpreter.CompositeStaticType{
 				Location:            TestLocation,
 				QualifiedIdentifier: "S",
-				TypeID:
 			},
 		},
 		{
